requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
beautifulsoup4==4.12.3
blinker==1.7.0
bootstrap4==0.1.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
cssselect2==0.7.0
dash==2.9.3
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
defusedxml==0.7.1
dj-database-url==2.2.0
Django==4.2.9
django-bootstrap-icons==0.8.7
django-classy-tags==4.1.0
django-cms==4.1.0
django-entangled==0.5.4
django-environ==0.11.2
django-filer==3.1.1
django-formtools==2.5.1
django-fsm==3.0.0
django-guest-user==0.5.5
django-livereload==1.7
django-parler==2.3
django-plotly-dash==2.3.1
django-polymorphic==3.1.0
django-sekizai==4.1.0
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
djangocms-admin-style==3.3.1
djangocms-alias==2.0.1
djangocms-attributes-field==3.0.0
djangocms-frontend==1.3.1
djangocms-text-ckeditor==5.1.5
djangocms-versioning==2.0.1
dpd-components==0.1.0
dpd-static-support==0.0.5
easy-thumbnails==2.8.5
Flask==3.0.3
fonttools==4.53.0
gunicorn==22.0.0
html5lib==1.1
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.2.1
MarkupSafe==2.1.5
oscrypto==1.3.0
packaging==24.0
pillow==10.3.0
plotly==5.20.0
psycopg2-binary==2.9.9
pycparser==2.22
pydyf==0.10.0
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
pypdf==4.2.0
pyphen==0.15.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dotenv==1.0.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
svglib==1.5.1
tenacity==8.2.3
tinycss2==1.2.1
typing_extensions==4.11.0
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.1
validate-email==1.3
weasyprint==61.2
webencodings==0.5.1
Werkzeug==3.0.2
whitenoise==6.6.0
xhtml2pdf==0.2.15
zipp==3.18.1
zopfli==0.2.3
